Reading LAI instead of LAIv
In this commit, I changed the LAIv reading to LAI. The code now reads the climatology of LAI from the Copernicus 1 km LAI dataset and calculates LAIv when running the canopy model. Two new namelist options are added to explicitly specify the number of LAI records per year and the scaling factor for the data.