A major modification towards regridding
The interpolation module for prep_megan changed to the old approach in MEGANv3.2. The new code is using the "Mode" and "Bilinear" approaches for interpolation, however, an "area conserving" approach is more suitable for MEGAN input datas including growthform inputs, ecotype, and LAIs. Therefore, I took back the old module in prepmegan4cmaq. In addition, following modifications are made:
    1. Correct the canopy tpye calculation.
    2. Change the way for organizing the emission factors, which are based on growthforms and ecotypes not canopy types.
    3. Cancel the inputs for "GRIDDESC" and "GRIDNAME". The code could read coordinates directly from the wrf file.
    4. Adding extra inputs including EFs and LDFs for each growthform in prep_megan_static.nc for diagnosis.
    5. src/src_ca4km could be used for code testing.